fix bug

+++ apps/api/src/features/master-data/domain/usecases/component/delete-component.usecase.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable, Logger } from '@nestjs/common';
+import { Inject, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { UseCase } from 'src/core/domain/usecases/usecase';
 import { DataState } from 'src/core/resources/data.state';
 import { COMPONENT_REPO_TOKEN } from 'src/core/const/provider.token';
@@ -16,6 +16,14 @@ export class DeleteComponentUsecase
   ) {}
 
   async execute(input: number): Promise<DataState<string>> {
+    this.logger.debug(`Checking component existence with id: ${input}`);
+    const existingComponent = await this.componentRepository.findById(input);
+
+    if (!existingComponent.data) {
+      this.logger.warn(`Component with id ${input} not found`);
+      throw new NotFoundException('Component not found');
+    }
+
     this.logger.debug(`Deleting component with id: ${input}`);
     const result = await this.componentRepository.delete(input);
